Logout Functionality

+++ app/components/CalendarHeader.jsx
@@ -4,10 +4,11 @@ import GlobalContext from "@/Context/GlobalContext";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
 import ChevronRightIcon from "@mui/icons-material/ChevronRight";
 import dayjs from "dayjs";
-import Image from "next/image";
 import Logo from "../../public/logo.png";
+import Image from "next/image";
 import AddIcon from "@mui/icons-material/Add";
 import DeleteIcon from "@mui/icons-material/Delete";
+import { useRouter } from "next/navigation";
 import fetchAllTimetables, {
   createTimetable,
   updateTimetable,
@@ -24,6 +25,8 @@ function CalendarHeader() {
     setCurrentTimetableIndex,
   } = useContext(GlobalContext);
   const [timetableName, setTimetableName] = useState("");
+  const router = useRouter();
+
   const handlePrevMonth = () => {
     setMonthIndex((prevState) => prevState - 1);
   };
@@ -55,8 +58,10 @@ function CalendarHeader() {
           timetables[currentTimetableIndex]._id,
           e.target.innerText
         );
-        dispatchTimetable({ type: "update", payload: res.data });
-        setTimetableName(e.target.innerText);
+        if (res) {
+          dispatchTimetable({ type: "update", payload: res.data });
+          setTimetableName(e.target.innerText);
+        }
       } else {
         setTimetableName((prevState) => prevState);
       }
@@ -80,33 +85,37 @@ function CalendarHeader() {
           )
         ) == JSON.stringify([])
       ) {
-        await createTimetable(`Demo ${timetables.length + 1}`)
-          .then((res) => {
-            setCurrentTimetableIndex(timetables.length);
-            console.log(res);
-            dispatchTimetable({ type: "push", payload: res.data });
-          })
-          .catch((err) => console.log(err));
+        const res = await createTimetable(`Demo ${timetables.length + 1}`);
+        if (res) {
+          setCurrentTimetableIndex(timetables.length);
+          console.log(res);
+          dispatchTimetable({ type: "push", payload: res.data });
+        }
       }
     })();
   };
   const timetableDeleteHandler = () => {
     (async () => {
-      await deleteTimetable(timetables[currentTimetableIndex]?._id)
-        .then((res) => {
-          if (currentTimetableIndex != 0)
-            setCurrentTimetableIndex(currentTimetableIndex - 1);
-          else {
-            setCurrentTimetableIndex(0);
-          }
-          console.log(res);
-          dispatchTimetable({
-            type: "delete",
-            payload: timetables[currentTimetableIndex]?._id,
-          });
-        })
-        .catch((err) => console.log(err));
+      const res = await deleteTimetable(timetables[currentTimetableIndex]?._id);
+      if (res) {
+        if (currentTimetableIndex != 0)
+          setCurrentTimetableIndex(currentTimetableIndex - 1);
+        else {
+          setCurrentTimetableIndex(0);
+        }
+        console.log(res);
+        dispatchTimetable({
+          type: "delete",
+          payload: timetables[currentTimetableIndex]?._id,
+        });
+      } else {
+        console.log("h");
+      }
     })();
+  };
+  const logoutHandler = () => {
+    localStorage.setItem("token", "");
+    router.push("/");
   };
   useEffect(() => {
     (async () => {
@@ -120,45 +129,31 @@ function CalendarHeader() {
       }
     })();
   }, []);
-  // useEffect(() => {
-  //   (async () => {
-  //     const tt = await fetchAllTimetables();
-  //     console.log(tt.data);
-  //     if (JSON.stringify(tt.data) == JSON.stringify([])) {
-  //       // console.log(tt.data);
-  //       const res = await createTimetable("Demo");
-  //       console.log(res.data);
 
-  //       if (res.data != undefined)
-  //         dispatchTimetable({ type: "push", payload: [res.data] });
-  //     } else {
-  //       setTimetableName(tt.data[0].name);
-  //       // console.log(tt.data[0].name);
-  //       if (!tt.data.includes(undefined))
-  //         dispatchTimetable({ type: "push", payload: [tt.data] });
-  //     }
-  //   })();
-  // }, []);
   return (
-    <header className="px-4 py-2 flex items-center">
-      <Image src={Logo} width={25} height={25} alt="Picture of the author" />
+    <div className="px-4 py-2 grid grid-cols-3 items-center  w-screen">
+      <div className="flex justify-center items-center">
+        <Image src={Logo} width={25} height={25} alt="Logo" className="mx-3" />
+        <button className="border rounded py-2 px-4 " onClick={handleReset}>
+          Today
+        </button>
 
-      <button className="border rounded py-2 px-4 mr-5" onClick={handleReset}>
-        Today
-      </button>
-      <button onClick={handlePrevMonth}>
-        <span>
-          <ChevronLeftIcon className="cursor-pointer text-gray-600 mx-2" />
-        </span>
-      </button>
-      <button onClick={handleNextMonth}>
-        <span className="cursor-pointer text-gray-600 mx-2">
-          <ChevronRightIcon />
-        </span>
-      </button>
-      {/* if month index is >11 it will move on to the next year */}
-      <h2>{dayjs(new Date(dayjs().year(), monthIndex)).format("MMMM YYYY")}</h2>
-      <div className="flex ml-10">
+        <button onClick={handlePrevMonth}>
+          <span>
+            <ChevronLeftIcon className="cursor-pointer text-gray-600 mx-2" />
+          </span>
+        </button>
+        {/* if month index is >11 it will move on to the next year */}
+        <h2 className="">
+          {dayjs(new Date(dayjs().year(), monthIndex)).format("MMMM YYYY")}
+        </h2>
+        <button onClick={handleNextMonth}>
+          <span className="cursor-pointer text-gray-600 mx-2">
+            <ChevronRightIcon />
+          </span>
+        </button>
+      </div>
+      <div className="flex mx-12">
         <button onClick={prevTimetableHandler}>
           <span>
             <ChevronLeftIcon className="cursor-pointer text-gray-600 mx-2" />
@@ -175,20 +170,25 @@ function CalendarHeader() {
           </span>
         </button>
       </div>
-      <button onClick={createTimetableHandler}>
-        <span className="cursor-pointer text-gray-600 mx-2">
-          <AddIcon />
-        </span>
-      </button>
-      <button
-        onClick={timetableDeleteHandler}
-        disabled={timetables?.length == 1}
-      >
-        <span className="cursor-pointer text-gray-600 mx-2">
-          <DeleteIcon />
-        </span>
-      </button>
-    </header>
+      <div className="mx-10">
+        <button onClick={createTimetableHandler}>
+          <span className="cursor-pointer text-gray-600 mx-2">
+            <AddIcon />
+          </span>
+        </button>
+        <button
+          onClick={timetableDeleteHandler}
+          disabled={timetables?.length == 1}
+        >
+          <span className="cursor-pointer text-gray-600">
+            <DeleteIcon />
+          </span>
+        </button>
+        <button onClick={logoutHandler} className="mx-10">
+          Logout
+        </button>
+      </div>
+    </div>
   );
 }
